requirements.txt
# Web Framework
streamlit==1.28.1
fastapi==0.104.1
uvicorn==0.24.0

# PDF and Document Processing
PyMuPDF==1.23.8
pdfplumber==0.9.0
python-docx==0.8.11
docx2txt==0.8

# NLP and Text Processing
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1

# AI/ML Libraries
openai==1.3.5
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.4
pandas==1.5.3

# LangChain Framework (Required)
langchain==0.1.20
langchain-openai==0.1.8
langchain-core==0.1.52
langgraph==0.0.69
langsmith==0.1.85

# Vector Search and Embeddings
faiss-cpu==1.7.4
chromadb==0.4.15

# Database
sqlite3  # Built into Python
psycopg2-binary==2.9.9  # For PostgreSQL if needed

# Utilities
python-dotenv==1.0.0
requests==2.31.0
regex==2023.10.3
tqdm==4.66.1

# Data Validation
pydantic==2.5.0

# File Upload and Processing
python-multipart==0.0.6
Pillow==10.1.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Fuzzy String Matching
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1

# Additional NLP Models (optional)
# en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
